+++ plone/app/blocks/tests/test_tiles.py
@@ -13,6 +13,13 @@ from zope.interface import Interface
 import unittest
 
 
+try:
+    # Python 2: "unicode" is built-in
+    unicode
+except NameError:
+    unicode = str
+
+
 class ITestTile(Interface):
 
     magicNumber = schema.Int(title=u"Magic number", required=False)
@@ -35,6 +42,7 @@ class TestTile(Tile):
   <p>
     Magic number: %(number)d; Form: %(form)s;
     Query string: %(queryString)s; URL: %(url)s
+    Umlauts: Übertile ;)
   </p>
 </body>
 </html>""" % dict(name=self.id, number=self.data['magicNumber'] or -1,
@@ -221,10 +229,11 @@ class TestRenderTiles(unittest.TestCase):
         request = self.layer['request']
         tree = serializer.tree
         renderTiles(request, tree)
-        result = str(serializer)
+        result = unicode(serializer)
         self.assertIn('This is a demo tile with id tile2', result)
         self.assertIn('This is a demo tile with id tile3', result)
         self.assertIn('This is a demo tile with id tile4', result)
+        self.assertIn(u'Umlauts: \xdcbertile', result)
 
     def testRenderTilesError(self):
         serializer = getHTMLSerializer([testLayout2])
@@ -234,7 +243,7 @@ class TestRenderTiles(unittest.TestCase):
         result = str(serializer)
         self.assertIn('This is a demo tile with id tile2', result)
         self.assertNotIn('This is a demo tile with id tile3', result)
-        self.assertIn('There was an error while rendering this tile', result)
+        self.assertIn('There was an error', result)
         self.assertIn('This is a demo tile with id tile4', result)
 
     def testRenderSubTiles(self):